Changes as per @TheDoubleK comments
* New file for CSV from API base class
* New arg name in script to bring more in line with other scripts
* Handling for other status types
